+++ config/routes.rb
@@ -1,8 +1,6 @@
 Rails.application.routes.draw do
 
   root 'posts#index'
-  get 'friendships/new'
-  get 'friendships/create'
 
   devise_for :users